Going Scriptable
Changed the ability system to be scriptableObject.
AnimatedEnemy class has been stripped down to the minimum.
EnemyAbility Recreated
Ab_HeavyAttack removed, as it was not scriptableObject
EnemyAbiilityManager is now serialized, meaning every GameObject having an instance of a EnemyAbilityManager will display it in the inspector. This is implemented so that the list<ability> in E.A.Manager  can be public.  Therefore, adding abilities to an Enemy is a drag and drop.